pyproject.toml
[tool.poetry]
name = "h2o-transcribe"
version = "1.0.0"
description = "H2O Transcribe is a general-purpose speech recognition webservice."
homepage  = "https://github.com/asedaghat/H2O_Transcribe"
authors = [
    "Amin Sedaghat",
]
readme = "README.md"
packages = [{ include = "app" }]

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
secondary = true

[tool.poetry.dependencies]
python = "^3.10"
unidecode = "^1.3.4"
uvicorn = { extras = ["standard"], version = "^0.18.2" }
gunicorn = "^20.1.0"
tqdm = "^4.64.1"
transformers = "^4.22.1"
python-multipart = "^0.0.5"
ffmpeg-python = "^0.2.0"
fastapi = "^0.95.1"
llvmlite = "^0.39.1"
numba = "^0.56.4"
openai-whisper = "20230124"
faster-whisper = "^0.4.1"
torch = [
  {markers = "sys_platform == 'darwin' and platform_machine == 'arm64'", url = "https://download.pytorch.org/whl/cpu/torch-1.13.0-cp310-none-macosx_11_0_arm64.whl"},
  {markers = "sys_platform == 'linux' and platform_machine == 'arm64'", url="https://download.pytorch.org/whl/cpu/torch-1.13.0-cp310-none-macosx_11_0_arm64.whl"},
  {markers = "sys_platform == 'darwin' and platform_machine == 'x86_64'", url = "https://download.pytorch.org/whl/cpu/torch-1.13.0-cp310-none-macosx_10_9_x86_64.whl"},
  {markers = "sys_platform == 'linux' and platform_machine == 'aarch64'", url="https://download.pytorch.org/whl/torch-1.13.0-cp310-cp310-manylinux2014_aarch64.whl"},
  {markers = "sys_platform == 'linux' and platform_machine == 'x86_64'", url="https://download.pytorch.org/whl/cpu/torch-1.13.0%2Bcpu-cp310-cp310-linux_x86_64.whl"},
  {markers = "sys_platform == 'win' and platform_machine == 'amd64'", url="https://download.pytorch.org/whl/cpu/torch-1.13.0%2Bcpu-cp310-cp310-win_amd64.whl"},
]

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
